+++ common/r_model.c
@@ -26,15 +26,20 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
  * Find the correct interval based on time
  * Used for Alias model frame/sprite/skin group animations
  */
-int Mod_FindInterval(const float *intervals, int numintervals, float time)
+int Mod_FindInterval(const float *intervals, int numintervals, double time)
 {
    int i;
    /*
     * when loading models/skins/sprites, we guaranteed all interval values
     * are positive, so we don't have to worry about division by 0
+    *
+    * Compute (time mod fullinterval) in double so cl.time-derived inputs
+    * keep full precision regardless of how long the engine has been
+    * running, then cast the bounded result (< fullinterval) to float for
+    * the comparison loop against the (small) interval table.
     */
    float fullinterval = intervals[numintervals - 1];
-   float targettime   = time - (int)(time / fullinterval) * fullinterval;
+   float targettime   = (float)(time - (int)(time / fullinterval) * fullinterval);
 
    for (i = 0; i < numintervals - 1; i++)
       if (intervals[i] > targettime)
